Update Product.php
When assigning products to a configurable product line 26 is removing any already assigned and adding the new product. 

If I have a bunch of products assigned I can not add another product to this using an update to the configurable product. If this line was changed to array_merge($oldProductIds,$newProductIds) this would solve that issue.